Stop using our customized version of Giphy4J
Giphy4J seemed like the quickest way to get things working, but it
was too simplistic and had to be modified to meet our needs.  Now we
can get rid of it and just use okhttp and org.json.  No need to carry
around a modified version of Giphy4J anymore.