Fix some problems that appeared on my Samsung S6
Get rid of TransactionTooLargeException by clearing the saved state
in onSaveInstanceState().

The SearchView focus works differently on the Samsung S6 than the
Nexus 5X.  When you close the soft keyboard on the 5X, the SearchView
still has focus, but on the S6 it looses focus.  So instead of using
the SearchView's focus state to determine if we are in search mode,
use the SearchView's menu item's expanded/collapsed state instead,
which seems to be what we want.