Get rid of "Error at position X" when changing query
If we are far down in the list, and we change the query (e.g. by
changing the rating), then

    Gif gif = getItem(position);

may return null.  In that case, we had been showing and error message
instead of the GIF.

But since that error message will confuse users, we remove it in
this commit.

Another possible solution would have been to scroll to the top every
time we change the query, but then we'd miss the cool animations that
Android does when moving things aroung.